test: redact sensitive data in test logs
