+++ e2e-tests/src/log_filter.py
@@ -1,0 +1,22 @@
+import logging
+import re
+
+
+class SensitiveDataFilter(logging.Filter):
+    def __init__(self, patterns):
+        super().__init__()
+        self.patterns = patterns
+
+    def filter(self, record):
+        message = record.getMessage()
+        for pattern, replacement in self.patterns:
+            message = re.sub(pattern, replacement, message)
+        record.msg = message
+        return True
+
+
+mc_vkey_pattern = (r"mc_vkey='([^']*)'", "mc_vkey='[REDACTED]'")
+signing_key_arg_pattern = (re.compile(r"(--signing-key\s+|--mainchain-signing-key\s+|--sidechain-signing-key\s+)[^\s]+", re.IGNORECASE), r"\1[REDACTED]") 
+signing_key_file_pattern = (re.compile(r"(SigningKey.*?cborHex.: .)([0-9a-fA-F]+)(.)", re.IGNORECASE | re.DOTALL), r"\1[REDACTED]\3")
+
+sensitive_filter = SensitiveDataFilter([mc_vkey_pattern, signing_key_arg_pattern, signing_key_file_pattern])
